+++ MainScan.cpp
@@ -6,7 +6,6 @@
 #include "Utils.h"
 #include "Reports.h"
 #include "Scanner.h"
-#include "MainScan.h"
 
 #include "Version.info"
 
@@ -16,6 +15,41 @@
 #define ARG(S) ARGX3(L"-" #S, L"--" #S, L"/" #S)
 #define ARGPARAMCOUNT(X) ((i + X) <= (argc - 1))
 
+
+class CCommandLineOptions {
+ public:
+  bool scanLocalDrives;
+  bool scanLocalDrivesInclMountpoints;
+  bool scanNetworkDrives;
+  bool scanFile;
+  std::wstring file;
+  bool scanDirectory;
+  std::wstring directory;
+  bool report;
+  bool reportPretty;
+  bool reportSig;
+  bool lowpriority;
+  bool verbose;
+  bool no_logo;
+  bool help;
+
+  CCommandLineOptions() {
+    scanLocalDrives = false;
+    scanLocalDrivesInclMountpoints = false;
+    scanNetworkDrives = false;
+    scanFile = false;
+    file.clear();
+    scanDirectory = false;
+    directory.clear();
+    report = false;
+    reportPretty = false;
+    reportSig = false;
+    lowpriority = false;
+    verbose = false;
+    no_logo = false;
+    help = false;
+  }
+};
 
 CCommandLineOptions cmdline_options;
